requirements.txt
# Core dependencies
jsonschema>=4.17.0
python-dateutil>=2.8.2
pytz>=2023.3
pydantic>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0